backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
supabase==2.15.1
python-dotenv==1.0.1
pydantic-settings==2.9.1
mem0ai==2.0.0
psycopg[binary,pool]==3.3.3
langfuse==4.5.0
celery==5.6.3
redis==6.4.0
pdfminer.six==20231228
python-multipart==0.0.20
